♻️ refactor(Category): 카테고리 컬럼에서 엔티티로 변경 및 기본 기능 추가 (#133)
