+++ src/main/java/com/sparta/devquiz/domain/category/enums/QuizCategory.java
@@ -1,4 +1,4 @@
-package com.sparta.devquiz.domain.quiz.enums;
+package com.sparta.devquiz.domain.category.enums;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
@@ -53,7 +53,6 @@ public enum QuizCategory {
     DOCKER("DOCKER"),
     GIT("GIT"),
     LINUX("LINUX");
-
 
     private final String major;
