--- src/main/java/com/sparta/devquiz/domain/quiz/repository/CategoryRepository.java
@@ -1,10 +1,0 @@
-package com.sparta.devquiz.domain.quiz.repository;
-
-import com.sparta.devquiz.domain.quiz.entity.Category;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface CategoryRepository extends JpaRepository<Category, Long> {
-        Optional<Category> findById(Long categoryId);
-}
